I used bitmask DP (soln continued in description).
For every m, a row was a bunch of bits, 1 for the top of a vertical (that will go to the next row), and 0 for anything else (that will only stay on the current row). This way, I could use the fact that all the 0s would not carry over to the next row to my advantage. I first calculated all of the possible next states for each current state, and even though this might initially seem to be 2^n, a lot of those states are eliminated, and it can be proved that the maximum amount of next states is the nth Fibonacci number (in practice, this ends up being an average of around 10-20 or so for this n). Overall, the complexity is O(m * 2^n * fib(n)), which is fast enough to pass.